Large Rework Progress
- Entity Register:
remastered how i generate new entities. they now use a bit based block to track which entities are active. there is no internal id storage just bool checks for active. fully modular and interfaces with archetype registry through the use of a Entity meta card.
- Component Register:
remaster component registration through this new class. all components are structs and cannot have reference types stored as properties, this is an enforced feature. components are assigned meta cards that are stored inside the register for lookups and quick informational overviews. the register is in charge of creating all component sets which are used inside archetypes for data storage indexed by entities.
- Component Set:
A complete rethink on how I store data for entities as well as its access patterns, optimized as best as I know how for CPU and stack efficiency(for what c# allows). stores a set of components in an array of arrays and requires T to be used when trying to access contents in it as well as a index(entity id).
- Component Memory<T>:
This is a major turning point for how i allow components to be accessed, edited, set and all other needs. this stores a full or slice of a component array as in the ref struct as a span of that memory location so it can be used easily by the CPU. the component set is in charge of allowing access through this struct via helper methods.
- Component Template:
A simple data template that stores a empty set of component arrays with some meta information to allow for fast creation of new sets in an archetype. it stores a mask, capacity, meta info of components and the arrays themselves. it has a clone method to allow for creating duplicates of itself for creating the new fresh sets with no corrupted data values.
- Chunk: (renamed old Chunk to Code Register)
This is a new class that stores a single component set and an array of entities. its main job is interfacing local entity ids to indexes of components for quick iteration in systems.
- Archetype:
Remastered the entire class to support chunk based expansion using a new class called Bit Indexer for really fast entity allocation into empty chunks. the only goal of an archetype is to keep chunks tight and know its own size for meta data usage. NOTE this is still WIP as of now and will see some more changes as its not where I want it nor know where it could be in terms of speed, i also still havent thought of a proper query strategy that will allow for instant grabbing of all chunks into a queue of some sort for systems. more info to come as i code.
- Archetype Registry:
the latest addition which has been in the works and think tank for the past week now. its duty is to track and store archetypes along with their meta cards and keep a log of which entities it is currently influencing. there really isnt too much to it other than it being the head piece of the trickle down hierarchy for chunks.

What still needs to be done?
Queries - not started
Systems - not started
Archetypes - 80%
ArchetypeReg - 60%
Benchmarks - not started
Tests - 50% entities and components with some archetypes

there is still alot more work to be done in this current remaster and of course its in a broken ish state but everything from archetypes to entities functions and it does build. i have tests setup for the 3 i mentions in the tests folder if you want to play around with what is currently implemented. keep in mind remove does not work on archetypes yet so you can only add.

ADDITIONS:
alongside the rework i also added some api friendly helper classes to make dev life a little easier for this ECS and they are as follows:

- Entity Editor
- Entity Builder
- ChunkMask Builder
- Prefab: with only 1 generic so far plans for up to 10
- BitIndexer: which is meant to try and replace heavy reliance on dictionaries and is a generic use encoder/decoder(thanks for telling me about those terms Chand :D). the thought behind it is simple but complex due to the use of bitwise operations. Message for more info.

some of these require dependency injection but thats intended and will stay due to the modular nature of my ECS.