Lots of changes: - Discovered unsafe code and started to experiement with buffer and self allocated arrays - Finished up some major ECS features - Decided to rework a few features to incorporate some unsafe code(this will be after I finish experiementation of pointers and data structures) - Started to theorycraft some wrappers for the unsafe data structures
Unsafe TODO's: (some will be generic and some will be non generic with a internal restrictor to prevent bad data being added)
*default flat array (C# Array but buffer internal)
*padded array for vectorization and simd viable components (basically vector array)
*stack (C# Stack clone but pointer)
*hashset (C# HashSet but with own algo)
*hashmap (C# Dict but own algo)
*queue (C# Queue but pointer)
*allocator (basically my way of garbage collecting)
*many arenas (for memory management will be used inside allocators)

-will start working with threads to see about making allocators a background process and unintrusive as possible and will probably try using thread capable code more often even though ive been trying to
program in a way that allows it anyway ill be keeping it more mindful from now on.

- There will most likely be more as I learn but for now I want to implement just these as they are what I mainly see myself using in my ECS

(wanted to add this note here, alot of what I am doing is for fun since I simply enjoy the act of coding and the experiementation behind it. this is for the YAGNI followers that may see this :D)